feat: connect sdk concepts
